Move libmodplug from 0.8.7 to 0.8.8.1.

+++ contrib/libmodplug/src/fastmix.cpp
@@ -1970,7 +1970,7 @@ DWORD MPPASMCALL X86_Convert32To32(LPVOID lp16, int *pBuffer, DWORD lSampleCount
 {
 	UINT i ;
 	int vumin = *lpMin, vumax = *lpMax;
-	signed long *p = (signed long *)lp16;
+	int32_t *p = (int32_t *)lp16;
 	
 	for ( i=0; i<lSampleCount; i++)
 	{
@@ -2336,7 +2336,7 @@ UINT MPPASMCALL X86_AGC(int *pBuffer, UINT nSamples, UINT nAGC)
 
 	while(nSamples)
 	{
-		x = ((long long int)(*pBuffer) * nAGC) >> AGC_PRECISION;
+		x = ((int64_t)(*pBuffer) * nAGC) >> AGC_PRECISION;
 
 		if((x < MIXING_LIMITMIN) || (x > MIXING_LIMITMAX))
 		nAGC--;
